refactor : 아티스트, 장르 삭제시 연관 키워드 엔티티 삭제
