+++ app/domain/show-domain/src/main/java/org/example/repository/artist/artistsearch/ArtistSearchRepository.java
@@ -1,10 +1,11 @@
 package org.example.repository.artist.artistsearch;
 
+import java.util.List;
 import java.util.UUID;
 import org.example.entity.artist.ArtistSearch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArtistSearchRepository extends JpaRepository<ArtistSearch, UUID>,
     ArtistSearchQuerydslRepository {
-
+    List<ArtistSearch> findAllByArtistIdAndIsDeletedFalse(UUID artistId);
 }
