+++ app/domain/show-domain/src/main/java/org/example/repository/show/showsearch/ShowSearchRepository.java
@@ -1,10 +1,11 @@
 package org.example.repository.show.showsearch;
 
+import java.util.List;
 import java.util.UUID;
 import org.example.entity.show.ShowSearch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ShowSearchRepository extends JpaRepository<ShowSearch, UUID>,
     ShowSearchQuerydslRepository {
-
+    List<ShowSearch> findAllByShowIdAndIsDeletedFalse(UUID showId);
 }
