+++ app/domain/show-domain/src/main/java/org/example/usecase/artist/ArtistUseCase.java
@@ -98,6 +98,9 @@ public class ArtistUseCase {
 
         List<ShowArtist> showArtists = showArtistRepository.findAllByArtistIdAndIsDeletedFalse(artist.getId());
         showArtists.forEach(BaseEntity::softDelete);
+
+        List<ArtistSearch> artistSearches = artistSearchRepository.findAllByArtistIdAndIsDeletedFalse(artist.getId());
+        artistSearches.forEach(BaseEntity::softDelete);
     }
 
     public ArtistDetailPaginationResponse searchArtist(
